fix: address lejepa review comments
