+++ lightly/loss/__init__.py
@@ -10,7 +10,7 @@ from lightly.loss.directclr_loss import DirectCLRLoss
 from lightly.loss.emp_ssl_loss import EMPSSLLoss
 from lightly.loss.ibot_loss import IBOTPatchLoss
 from lightly.loss.koleo_loss import KoLeoLoss
-from lightly.loss.lejepa_loss import SIGReg
+from lightly.loss.lejepa_loss import LeJEPALoss, SIGReg
 from lightly.loss.macl_loss import MACLLoss
 from lightly.loss.mmcr_loss import MMCRLoss
 from lightly.loss.msn_loss import MSNLoss
