+++ lightly/models/modules/__init__.py
@@ -29,7 +29,7 @@ from lightly.models.modules.heads import (
     SwaVProjectionHead,
     SwaVPrototypes,
 )
-from lightly.models.modules.lejepa import LeJEPAEncoder, LeJEPALoss
+from lightly.models.modules.lejepa import LeJEPAEncoder
 from lightly.models.modules.nn_memory_bank import NNMemoryBankModule
 from lightly.utils import dependency as _dependency
